+++ AdventOfCode/ProceduralOptimizedSolution.hpp
@@ -7,8 +7,8 @@
 namespace OptimizedProceduralSolution
 {
 	std::vector<std::vector<short>> ReadFile (const std::string& fileName);
-	int GetNumberOfVisibleTreesInForest (const std::vector<std::vector<short>>& forest);
-	int GetHighestScenicScoreInForest (const std::vector<std::vector<short>>& forest);
+	uint64_t GetNumberOfVisibleTreesInForest (const std::vector<std::vector<short>>& forest);
+	uint64_t GetHighestScenicScoreInForest (const std::vector<std::vector<short>>& forest);
 
 	void RunOptimizedProceduralSolution (const std::string& fileName);
 }
