+++ AdventOfCode/TestRunner.hpp
@@ -7,13 +7,13 @@ namespace Test
 {
 	struct TestConfig {
 		const std::string fileName;
-		const unsigned int numberOfTestRuns;
+		const size_t numberOfTestRuns;
 		const bool shouldRunFullTest;
 		const bool shouldMeasureFileRead;
 		const bool shouldRunVisibilityCountTest;
 		const bool shouldRunHighestScenicScoreTest;
 
-		TestConfig (const std::string& fileName, unsigned int numberOfTestRuns,
+		TestConfig (const std::string& fileName, size_t numberOfTestRuns,
 					bool shouldRunFullTest, bool shouldMeasureFileRead,
 					bool shouldRunVisibilityCountTest, bool shouldRunHighestScenicScoreTest);
 	};
